Move `gardensystem` component to `gardensystem/runtime`
